requirements.txt
robotframework
pydicom
requests